fixup! Add unit tests for project report detail view
