+++ employees/tests/test_views.py
@@ -66,3 +66,42 @@ class AdminReportViewTests(TestCase):
         self.assertEqual(self.report.description, self.data["description"])
         self.assertEqual(self.report.author, self.user)
         self.assertTrue(self.report.editable)
+
+
+class ProjectReportDetailTests(TestCase):
+    def setUp(self):
+        super().setUp()
+        self.user = UserFactory()
+        self.client.force_login(self.user)
+        self.report = ReportFactory()
+        self.url = reverse("project-report-detail", args=(self.report.pk,))
+        self.data = {
+            "date": self.report.date,
+            "description": self.report.description,
+            "project": self.report.project.pk,
+            "author": self.report.author.pk,
+            "task_activities": self.report.task_activities.pk,
+            "work_hours": self.report.work_hours,
+        }
+
+    def test_project_report_detail_view_should_display_report_details(self):
+        response = self.client.get(path=self.url)
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, self.report.description)
+        self.assertEqual(response.context_data["form"].instance, self.report)
+
+    def test_project_report_detail_view_should_update_report_on_post(self):
+        self.data['description'] = "Some other description"
+        response = self.client.post(path=self.url, data=self.data)
+        self.report.refresh_from_db()
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(self.report.description, self.data['description'])
+        self.assertTrue(self.report.editable)
+
+    def test_project_report_detail_view_should_not_update_report_on_post_if_form_is_invalid(self):
+        self.data['description'] = ''
+        response = self.client.post(path=self.url, data=self.data)
+        self.report.refresh_from_db()
+        self.assertEqual(response.status_code, 200)
+        self.assertIsNotNone(response.context_data["form"]._errors)
+        self.assertTrue(self.report.editable)
